auth reviews api's completed

+++ controllers/bootcamps.js
@@ -37,6 +37,21 @@ exports.getBootcamp = async (req, res, next) => {
 // @access  Private
 exports.createBootcamp = async (req, res, next) => {
   try {
+    // Add user to req.body
+    req.body.user = req.user.id;
+
+    // check for published bootcamp
+    const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+    // If the user is not admin they can only add one bootcamp
+    if (publishedBootcamp && req.user.role !== 'admin') {
+      return next(
+        new ErrorResponse(
+          `The user with ID ${req.user.id} has already published a bootcamp`,
+          400
+        )
+      );
+    }
     const data = await Bootcamp.create(req.body);
     res.status(201).json({
       success: true,
@@ -53,10 +68,7 @@ exports.createBootcamp = async (req, res, next) => {
 // @access  Private
 exports.updateBootcamp = async (req, res, next) => {
   try {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true,
-    });
+    let bootcamp = await Bootcamp.findById(req.params.id);
     if (!bootcamp) {
       // return res.status(400).json({ success: false });
       return next(
@@ -64,6 +76,20 @@ exports.updateBootcamp = async (req, res, next) => {
       );
     }
 
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+      return next(
+        new ErrorResponse(
+          `User ${req.params.id} is not authorized to update this bootcamp`,
+          401
+        )
+      );
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+      runValidators: true,
+    });
     res.status(200).json({ success: true, data: bootcamp });
   } catch (error) {
     // res.status(400).json({ success: false });
@@ -81,6 +107,15 @@ exports.deleteBootcamp = async (req, res, next) => {
       // return res.status(400).json({ success: false });
       return next(
         new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+      );
+    }
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+      return next(
+        new ErrorResponse(
+          `User ${req.params.id} is not authorized to delete this bootcamp`,
+          401
+        )
       );
     }
     await bootcamp.deleteOne();
@@ -133,6 +168,17 @@ exports.bootcampPhotoUpload = async (req, res, next) => {
         new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
       );
     }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+      return next(
+        new ErrorResponse(
+          `User ${req.params.id} is not authorized to update this bootcamp`,
+          401
+        )
+      );
+    }
+
     if (!req.files) {
       return next(new ErrorResponse(`please upload a file `, 400));
     }
