+++ controllers/courses.js
@@ -1,6 +1,6 @@
-const ErrorResponse = require('../utils/errorResponse');
-const Course = require('../models/Course');
-const Bootcamp = require('../models/Bootcamp');
+const ErrorResponse = require("../utils/errorResponse");
+const Course = require("../models/Course");
+const Bootcamp = require("../models/Bootcamp");
 
 // @desc   Get Courses
 // @route  GET /api/v1/courses
@@ -31,8 +31,8 @@ exports.getCourses = async (req, res, next) => {
 exports.getSingleCourse = async (req, res, next) => {
   try {
     const course = await Course.findById(req.params.id).populate({
-      path: 'bootcamp',
-      select: 'name description',
+      path: "bootcamp",
+      select: "name description",
     });
 
     if (!course) {
@@ -55,6 +55,7 @@ exports.getSingleCourse = async (req, res, next) => {
 exports.addCourse = async (req, res, next) => {
   try {
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
     if (!bootcamp) {
@@ -63,6 +64,16 @@ exports.addCourse = async (req, res, next) => {
           `No Bootcamp find with the id of ${req.params.bootcampId}`
         ),
         404
+      );
+    }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to add a course to this bootcamp ${bootcamp._id} `,
+          401
+        )
       );
     }
 
@@ -90,6 +101,16 @@ exports.updateCourse = async (req, res, next) => {
       );
     }
 
+    // Make sure user is course owner
+    if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to update course ${course._id}`,
+          401
+        )
+      );
+    }
+
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
       runValidators: true,
@@ -103,8 +124,8 @@ exports.updateCourse = async (req, res, next) => {
   }
 };
 
-// @desc   Update a Course
-// @route  PUT /api/v1/courses/:id
+// @desc   Delete a Course
+// @route  Delete /api/v1/courses/:id
 // @access Private
 
 exports.deleteCourse = async (req, res, next) => {
@@ -116,6 +137,17 @@ exports.deleteCourse = async (req, res, next) => {
         404
       );
     }
+
+    // Make sure user is course owner
+    if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to delete course ${course._id}`,
+          401
+        )
+      );
+    }
+
     await course.deleteOne();
     res.status(200).json({
       success: true,
